Use macro to store correct testcase name
Since check uses macros to get the testcase name, and the tests are added from an array, the error looks like this:
../../../../lwip/test/unit/core/test_pbuf.c:49:E:Core:tests[i]:0: (after this point) Received signal 6 (Aborted)
(Note the tests[i] part)

This patch creates a new macro used when listing test functions in the array, which stores the name earlier.
Then the test_add macro is modified to use the correct test name. New error:
../../../../lwip/test/unit/core/test_pbuf.c:49:E:PBUF:test_pbuf_copy_zero_pbuf:0: (after this point) Received signal 6 (Aborted)

This makes it easier to find the name of the failing test function.